Nutritionist_GenAI_Doctor/requirements.txt
streamlit 
google-genai 
pillow
python-dotenv